fixup! Use PyInquirer
